+++ src/joybus/libdragon.mk
@@ -1,0 +1,14 @@
+
+LIBDRAGON_OBJS += \
+	$(BUILD_DIR)/joybus/controller.o \
+	$(BUILD_DIR)/joybus/cpak.o \
+	$(BUILD_DIR)/joybus/eeprom.o \
+	$(BUILD_DIR)/joybus/eepromfs.o \
+	$(BUILD_DIR)/joybus/joybus_accessory.o \
+	$(BUILD_DIR)/joybus/joybus_rtc.o \
+	$(BUILD_DIR)/joybus/joybus.o \
+	$(BUILD_DIR)/joybus/joypad_accessory.o \
+	$(BUILD_DIR)/joybus/joypad.o \
+	$(BUILD_DIR)/joybus/mempak.o \
+	$(BUILD_DIR)/joybus/pixelfx.o \
+	$(BUILD_DIR)/joybus/tpak.o
