+++ src/rtc.c
@@ -7,7 +7,7 @@
 #include "bb_rtc.h"
 #include "dd.h"
 #include "debug.h"
-#include "joybus_rtc.h"
+#include "joybus/joybus_rtc.h"
 #include "libcart/cart.h"
 #include "n64sys.h"
 #include "rtc.h"
